pyproject.toml
[project]
name = "naive-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-qdrant>=0.2.0",
    "logging>=0.4.9.6",
    "ollama>=0.5.3",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "youtube-transcript-api>=1.2.2",
    "yt-dlp>=2025.8.11",
]
